+++ src/types/list.rs
@@ -213,7 +213,7 @@ pub struct ListsType {
     pub lists: Vec<ListType>,
     /// Desc: The total number of items matching the query regardless of pagination.
     #[serde(default)]
-    pub total_items: u32,
+    pub total_items: u64,
     /// Desc: A list of link types and descriptions for the API schema documents.
     #[serde(default)]
     pub _links: Vec<LinkType>,
@@ -221,7 +221,7 @@ pub struct ListsType {
 
 impl MailchimpCollection<ListType> for ListsType {
     /// Total Items
-    fn get_total_items(&self) -> u32 {
+    fn get_total_items(&self) -> u64 {
         self.total_items
     }
     /// Data
